0_project_3_about_tasty_and_healthy_food
As a Data Scientist of TripAdvisor, I need to analyse data available about restaurants and figure out which restaurant is cheating its ratings. The goal is to create a model which will determine a potential rating of a restaurant. If the model gives a rating that is very different from teh actual rating of the restaurant, it will show the restaurant is cheating its ratings.